First pass at reorganizing/cleaning up the MIDI part.
A few other files were touched as well as they were referred to by files
in the MIDI part, which lead to other connected things being pulled in.
As this is only the first pass, there will be more necessary changes
coming in the near future.  Also, the lowest two levels of structure
were mistakenly removed from the build system; these have been added
back as they are necessary to keep the structure of the manual sane.