Clarifying in which mode changes can be edited
The above paragraph suggests that all changes are done in the Grab Mode instead of the Internal Edit Mode. It isn't intuitive to change the mode from G to E to edit the patch changes in view of the documentation and several users ran into this problem:
https://discourse.ardour.org/t/how-do-i-modify-delete-a-patch-change/88413
https://discourse.ardour.org/t/g-e-mousemode-colors/89766

The documentation is clarified that while insertion happens in G mode, modification hapens in E mode.